Adhering to project structure best practices
