--- copis.ini
@@ -1,6 +1,0 @@
-[General]
-windowwidth = 1400
-windowheight = 1000
-
-[Test]
-test = ''
