+++ copis/copis.ini
@@ -1,0 +1,6 @@
+[General]
+windowwidth = 1400
+windowheight = 1000
+
+[Test]
+test = ''
