+++ copis/copiscore.py
