+++ copis/enums.py
