+++ copis/gui/panels/console.py
