+++ copis/gui/panels/toolbar.py
