+++ copis/gui/panels/visualizer.py
