+++ copis/gui/pathgen_frame.py
