+++ copis/gui/settings_frame.py
