+++ copis/util/Canon/CameraModel.py
