+++ copis/util/edsdk_object.py
